Add files via upload
This analysis was focused on the 2016 Presidential election. I merged two datasets. True.csv was on verified and true articles in the lead-up to the 2016 election, and Fake.csv was news that was unverifiable and fake and went viral in the lead-up to the 2016 elections. I cleaned the data, visualized and did a sentiment analysis on the data